+++ changelog.md
@@ -2,12 +2,15 @@ Change Log
 ==========
 
 
-v0.3 - 2014-xx-xx
+v0.3 - 2014-03-02
 -----------------
 
+* Added a way to import/save user's own tables. Right now you it either accepts a full JSON object or a simple one entry per line text format.
 * Added code documentation with JSDoc3 commenting.
 * Cleaned up RandomTable class and AppRandomizer functioning to be, hopefully, simpler and more consistent.
 * Moved more of the pre-RandomTable stuff into the RandomTable format/class (like NPC personalities, appearance, goals, and occupations).
+* More wilderness encounter tables (mostly in-process).
+* Added the "list table" function for complex tables. It should show all the tables and subtables now in a semi-usable format. (Depending on the weight values, it might not convert to real dice rolls.)
 
 
 v0.2 - 2014-02-24